fix package name - move `search.soft.gotodef` to `search.gotodef.soft`
